requirements.txt
cx-Freeze==4.3.4
numpy==1.11.0
pandas==0.18.1
pyserial==3.0.1
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
XlsxWriter==0.8.4
